requirements.txt
django
psycopg2-binary
requests
plotly
pandas
numpy